data-pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "predictions-ingest"
version = "2.0.0"
description = "Multi-source prediction market data ingestion system (Polymarket, Kalshi, Limitless)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Data Engineering Team"}
]
keywords = ["predictions", "polymarket", "kalshi", "limitless", "data", "ingestion", "etl", "prediction-markets"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy[asyncio]>=2.0.25",
    "alembic>=1.13.1",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "click>=8.1.7",
    "rich>=13.7.0",
    "tabulate>=0.9.0",
    "apscheduler>=3.10.4",
    "tenacity>=8.2.3",
    "structlog>=24.1.0",
    "orjson>=3.9.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.1.14",
    "black>=24.1.0",
    "types-pyyaml>=6.0.12",
    "types-tabulate>=0.9.0",
]

[project.scripts]
predictions-ingest = "predictions_ingest.cli:main"
# Legacy alias
limitless-ingest = "limitless_ingest.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["predictions_ingest*", "limitless_ingest*"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
